requirements.txt
anthropic>=0.40.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
streamlit>=1.35.0
pdfplumber>=0.11.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.13.0
python-multipart>=0.0.9
requests>=2.32.0
pydantic>=2.7.0
python-dotenv>=1.0.0
